[scsi] scsi: lpfc: Fix FLOGI handling across multiple link up/down conditions
Message-id: <20191001204158.21024-26-dkennedy@redhat.com>
Patchwork-id: 275917
O-Subject: [RHEL-8.2 PATCH e-stor 25/78] scsi: lpfc: Fix FLOGI handling across multiple link up/down conditions
Bugzilla: 1757574
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: James Smart <jsmart2021@gmail.com>

It's possible for the driver to initiate an FLOGI and before it completes,
another link down/up transition occurs requiring a new FLOGI. Currently,
nothing is done to abort/noop the older FLOGI request to the adapter, so if
this transition occurs and the FLOGI completion is received after the link
down/up transition, the driver may erroneously act on the older FLOGI. In
most cases, the adapter properly terminates/fails the FLOGI, but there is a
timing condition where the FLOGI may complete on the wire prior to the
transition, but the response may not be seen/processed by the driver before
the driver sees the link transition.

Fix by having the link down handler in the driver run through any
outstanding ELS's and change the completion handler of the ELS so that it
will be no-op'd and released.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 6ede2ddd8bdd28bdaf43d15e70dc0ef38c5448db)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>