[scsi] scsi: lpfc: Fix nvme target mode ABTSing a received ABTS
Message-id: <20191001204158.21024-38-dkennedy@redhat.com>
Patchwork-id: 275933
O-Subject: [RHEL-8.2 PATCH e-stor 37/78] scsi: lpfc: Fix nvme target mode ABTSing a received ABTS
Bugzilla: 1757574
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: James Smart <jsmart2021@gmail.com>

If an unsolicited ABTS was received, the driver looks up the exchange it
references. It it does various searches looking for the exchange
context. When one is eventually matched and it is associated with an XRI
context, the driver sends an ABORT WQE to terminate the exchange. Current
code looks at whether the transport had taken action on the XRI yet or not
(no action if set to LPFC_NVMET_STE_RCV; action if non-LPFC_NVMET_STE_RCV).
Based on action or not one of two (sol vs unsol) issue abort routines are
called. The unsol version cheats and transmits a sequence containing an
ABTS with no interaction with the adapter.  The sol version issues an Abort
WQE and lets the adapter manage whether the ABTS is sent to not.

The issue is the unsol version is sending ABTS unconditionally for the
exchange that received the ABTS. It's unnecessary.

Remove the conditional and just call the adapter command-based routine to
let the adapter manage the ABTS.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 005d8eb92847e6823d9fea6041f7c6f26b6cf2da)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>