[scsi] scsi: lpfc: Fix reset recovery paths that are not recovering
Message-id: <20191001204158.21024-59-dkennedy@redhat.com>
Patchwork-id: 275950
O-Subject: [RHEL-8.2 PATCH e-stor 58/78] scsi: lpfc: Fix reset recovery paths that are not recovering
Bugzilla: 1757574
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: James Smart <jsmart2021@gmail.com>

A recent patch unconditionally marks the hba as in error as part of
resetting the adapter. The driver flow that called the adapter reset was a
recovery path, which expects the adapter to not be in an error state in
order to finish the recovery.  Given the new error state being set, the
recovery fails and the adapter is left in limbo.

Revise the adapter reset routine so that it will only mark the adapter in
error if it was unable to reset the adapter.

Fixes: 8c24a4f643ed ("scsi: lpfc: Fix crash due to port reset racing vs adapter error handling")
Link: https://lore.kernel.org/r/20190903215441.10490-1-jsmart2021@gmail.com
Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 4fb86a6bc549fc482bdb21086aa272668ed955d2)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>