[scsi] scsi: lpfc: Fix NVMe ABTS in response to receiving an ABTS
Message-id: <20191001204158.21024-69-dkennedy@redhat.com>
Patchwork-id: 275953
O-Subject: [RHEL-8.2 PATCH e-stor 68/78] scsi: lpfc: Fix NVMe ABTS in response to receiving an ABTS
Bugzilla: 1757574
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: James Smart <jsmart2021@gmail.com>

When the port, running as a nvme target, receives an ABTS, it submits
commands to the adapter to Abort i/o outstanding in the adapter. The Abort
command formatting routine left a command field set to zero, which
instructs the adapter to generate an ABTS on the wire as part of cleaning
up the I/O. This is common operation for an initiator, but not for a
target.

Fix the driver to check whether an ABTS had been received for the I/O, and
if so, change the Abort command formatting so that the ABTS generation is
disabled (IA=1). No need to ABTS it when the other side already has.

Also refactored the code such that there is a single routine being used for
nvme or nvmet ABORT requests, and IA is an argument.

Link: https://lore.kernel.org/r/20190922035906.10977-11-jsmart2021@gmail.com
Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 51f8e43ed355d30b3c93293077ecb0c0afac3799)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>