[powerpc] powerpc/xmon: Fix output of XIVE IPI
Message-id: <20191002142358.13349-5-didoming@redhat.com>
Patchwork-id: 276024
O-Subject: [RHEL8.2 BZ1731130 PATCH 4/5] powerpc/xmon: Fix output of XIVE IPI
Bugzilla: 1731130
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1731130
Brew:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=23584534

When dumping the XIVE state of an CPU IPI, xmon does not check if the
CPU is started or not which can cause an error. Add a check for that
and change the output to be on one line just as the XIVE interrupts of
the machine.

commit 855d9140a394229f21fd4fd216f377ed45bd93a3

Signed-off-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20190910081850.26038-3-clg@kaod.org
Signed-off-by: Phillip Lougher <plougher@redhat.com>