[fs] cifs: fix smb3_zero_range for Azure
Message-id: <20191003012831.22506-21-lsahlber@redhat.com>
Patchwork-id: 276578
O-Subject: [RHEL8.2 BZ 1757670 20/79] cifs: fix smb3_zero_range for Azure
Bugzilla: 1757670
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Bugzilla: 1757670
Build: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=23812498

For zero-range that also extend the file we were sending this as a
compound of two different operations; a fsctl to set-zero-data for the range
and then an additional set-info to extend the file size.
This does not work for Azure since it does not support this fsctl which leads
to fallocate(FALLOC_FL_ZERO_RANGE) failing but still changing the file size.

To fix this we un-compound this and send these two operations as separate
commands, firsat one command to set-zero-data for the range and it this
was successful we proceed to send a set-info to update the file size.

This fixes xfstest generic/469 for Azure servers.

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
(cherry picked from commit c425014afd7f622d586e7242f94019131790f5a3)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>