[x86] kvm, x86: Properly check whether a pfn is an MMIO or not
Message-id: <20191010145416.6461-1-vkuznets@redhat.com>
Patchwork-id: 277911
O-Subject: [RHEL8.2 virt PATCH v3 042/330] kvm, x86: Properly check whether a pfn is an MMIO or not
Bugzilla: 1749495
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

From: KarimAllah Ahmed <karahmed@amazon.de>

pfn_valid check is not sufficient because it only checks if a page has a struct
page or not, if "mem=" was passed to the kernel some valid pages won't have a
struct page. This means that if guests were assigned valid memory that lies
after the mem= boundary it will be passed uncached to the guest no matter what
the guest caching attributes are for this memory.

Introduce a new function e820__mapped_raw_any which is equivalent to
e820__mapped_any but uses the original e820 unmodified and use it to
identify real *RAM*.

Signed-off-by: KarimAllah Ahmed <karahmed@amazon.de>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 0c55671f84fffe591e8435c93a8c83286fd6b8eb)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>