[x86] kvm: x86/pmu: do not mask the value that is written to fixed PMUs
Message-id: <20191010145737.7716-1-vkuznets@redhat.com>
Patchwork-id: 277937
O-Subject: [RHEL8.2 virt PATCH v3 068/330] KVM: x86/pmu: do not mask the value that is written to fixed PMUs
Bugzilla: 1749495
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

According to the SDM, for MSR_IA32_PERFCTR0/1 "the lower-order 32 bits of
each MSR may be written with any value, and the high-order 8 bits are
sign-extended according to the value of bit 31", but the fixed counters
in real hardware are limited to the width of the fixed counters ("bits
beyond the width of the fixed-function counter are reserved and must be
written as zeros").  Fix KVM to do the same.

Reported-by: Nadav Amit <nadav.amit@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 2924b52117b2812e9633d5ea337333299166d373)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>