[x86] kvm: nvmx: Ignore segment base for VMX memory operand when segment not FS or GS
Message-id: <20191010151416.14567-1-vkuznets@redhat.com>
Patchwork-id: 278066
O-Subject: [RHEL8.2 virt PATCH v3 185/330] KVM: nVMX: Ignore segment base for VMX memory operand when segment not FS or GS
Bugzilla: 1749495
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

From: Liran Alon <liran.alon@oracle.com>

As reported by Maxime at
https://bugzilla.kernel.org/show_bug.cgi?id=204175:

In vmx/nested.c::get_vmx_mem_address(), when the guest runs in long mode,
the base address of the memory operand is computed with a simple:
    *ret = s.base + off;

This is incorrect, the base applies only to FS and GS, not to the others.
Because of that, if the guest uses a VMX instruction based on DS and has
a DS.base that is non-zero, KVM wrongfully adds the base to the
resulting address.

Reported-by: Maxime Villard <max@m00nbsd.net>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Liran Alon <liran.alon@oracle.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 6694e48012826351036fd10fc506ca880023e25f)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>