[netdrv] ice: Increase size of Mailbox receive queue for many VFs
Message-id: <48591d62193b35280f897d4273d4e444df0e3196.1570730072.git.jtoppins@redhat.com>
Patchwork-id: 278307
O-Subject: [PATCH RHEL-8.2 074/143] ice: Increase size of Mailbox receive queue for many VFs
Bugzilla: 1721711
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Currently we use the ICE_MBXQ_LEN for both the Mailbox send and receive
queues that are used to communicate with VFs. This is fine for the send
queue because the PF driver will lock the queue for every single send,
but for the Mailbox receive queue every VF is posting to its Mailbox
send queue and the hardware is then handing the message to the PF on its
Mailbox receive queue. This becomes a problem with many VFs because it
seems to overburden the Mailbox receive queue on the PF. Fix this by
increasing the Mailbox receive queue for the PF to 512 entries.

The number 512 was determined based on the number of VFs supported by
the device. We can have a total of 256 VFs so in the worst case this
allows the VFs to put 2 messages in the PFs Mailbox receive queue at the
same time.

Signed-off-by: Brett Creeley <brett.creeley@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 11836214d5b7fbfe9df8aeb5dc2b23a6bedf185c)
Bugzilla: 1721711
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=23844481
Tested: build only, waiting on intel to test - no hardware
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>