[scsi] scsi: sd: Ignore a failure to sync cache due to lack of authorization
Message-id: <20191016164742.28086-54-emilne@redhat.com>
Patchwork-id: 279431
O-Subject: [RHEL8.2 PATCH v2 53/54] scsi: sd: Ignore a failure to sync cache due to lack of authorization
Bugzilla: 1760503
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

I've got a report about a UAS drive enclosure reporting back Sense: Logical
unit access not authorized if the drive it holds is password protected.
While the drive is obviously unusable in that state as a mass storage
device, it still exists as a sd device and when the system is asked to
perform a suspend of the drive, it will be sent a SYNCHRONIZE CACHE. If
that fails due to password protection, the error must be ignored.

Cc: <stable@vger.kernel.org>
Link: https://lore.kernel.org/r/20190903101840.16483-1-oneukum@suse.com
Signed-off-by: Oliver Neukum <oneukum@suse.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 21e3d6c81179bbdfa279efc8de456c34b814cfd2)
Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>