[misc] ocxl: Fix access to the AFU Descriptor Data
Message-id: <20191018131345.13133.1123.sendpatchset@stevebest>
Patchwork-id: 280099
O-Subject: [PATCH RHEL8.2 BZ1751207 01/22] ocxl: Fix access to the AFU Descriptor Data
Bugzilla: 1751207
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1751207

Description:
------------
The AFU Information DVSEC capability is a means to extract common,
general information about all of the AFUs associated with a Function
independent of the specific functionality that each AFU provides.
Write in the AFU Index field allows to access to the descriptor data
for each AFU.

With the current code, we are not able to access to these specific data
when the index >= 1 because we are writing to the wrong location.
All requests to the data of each AFU are pointing to those of the AFU 0,
which could have impacts when using a card with more than one AFU per
function.

This patch fixes the access to the AFU Descriptor Data indexed by the
AFU Info Index field.

Fixes: 5ef3166e8a32 ("ocxl: Driver code for 'generic' opencapi devices")
Cc: stable <stable@vger.kernel.org>     # 4.16
Signed-off-by: Christophe Lombard <clombard@linux.vnet.ibm.com>

Acked-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>
Acked-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 6f8e45f7eb1bee5efdbe4a9cfe4a45627403c5fb

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=6f8e45f7eb1bee5efdbe4a9cfe4a45627403c5fb

----------------
Steve Best

Proposed Patch:
---------------
 drivers/misc/ocxl/config.c | 4 +++-
 1 file changed, 3 insertions(+), 1 deletion(-)

Signed-off-by: Phillip Lougher <plougher@redhat.com>