[misc] ocxl: Fix endiannes bug in ocxl_link_update_pe()
Message-id: <20191018131355.13133.28379.sendpatchset@stevebest>
Patchwork-id: 280100
O-Subject: [PATCH RHEL8.2 BZ1751207 02/22] ocxl: Fix endiannes bug in ocxl_link_update_pe()
Bugzilla: 1751207
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1751207

Description:
------------
All fields in the PE are big-endian. Use cpu_to_be32() like everywhere
else something is written to the PE. Otherwise a wrong TID will be used
by the NPU. If this TID happens to point to an existing thread sharing
the same mm, it could be woken up by error. This is highly improbable
though. The likely outcome of this is the NPU not finding the target
thread and forcing the AFU into sending an interrupt, which userspace
is supposed to handle anyway.

Fixes: e948e06fc63a ("ocxl: Expose the thread_id needed for wait on POWER9")
Cc: stable@vger.kernel.org      # v4.18
Signed-off-by: Greg Kurz <groug@kaod.org>
Acked-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit e1e71e201703500f708bdeaf64660a2a178cb6a0

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=e1e71e201703500f708bdeaf64660a2a178cb6a0

----------------
Steve Best

Proposed Patch:
---------------
 drivers/misc/ocxl/link.c | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

Signed-off-by: Phillip Lougher <plougher@redhat.com>