[misc] ocxl: Fix endiannes bug in read_afu_name()
Message-id: <20191018131433.13133.15216.sendpatchset@stevebest>
Patchwork-id: 280093
O-Subject: [PATCH RHEL8.2 BZ1751207 05/22] ocxl: Fix endiannes bug in read_afu_name()
Bugzilla: 1751207
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1751207

Description:
------------
The AFU Descriptor Template in the PCI config space has a Name Space
field which is a 24 Byte ASCII character string of descriptive name
space for the AFU. The OCXL driver read the string four characters at
a time with pci_read_config_dword().

This optimization is valid on a little-endian system since this is PCI,
but a big-endian system ends up with each subset of four characters in
reverse order.

This could be fixed by switching to read characters one by one. Another
option is to swap the bytes if we're big-endian.

Go for the latter with le32_to_cpu().

Cc: stable@vger.kernel.org      # v4.16
Signed-off-by: Greg Kurz <groug@kaod.org>
Acked-by: Frederic Barrat <fbarrat@linux.ibm.com>
Acked-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 2f07229f02d4c55affccd11a61af4fd4b94dc436

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=2f07229f02d4c55affccd11a61af4fd4b94dc436

----------------
Steve Best

Proposed Patch:
---------------
 drivers/misc/ocxl/config.c | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

Signed-off-by: Phillip Lougher <plougher@redhat.com>