[fs] fuse: change interrupt requests allocation algorithm
Message-id: <20191029191438.6931-12-vgoyal@redhat.com>
Patchwork-id: 281730
O-Subject: [RHEL8.2 PATCH v2 011/134] fuse: change interrupt requests allocation algorithm
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit c59fd85e4fd07fdf0ab523a5e9734f5338d6aa19
Author: Kirill Tkhai <ktkhai@virtuozzo.com>
Date: Tue, 11 Sep 2018 13:11:56 +0300

Using of two unconnected IDs req->in.h.unique and req->intr_unique does not
allow to link requests to a hash table. We need can't use none of them as a
key to calculate hash.

This patch changes the algorithm of allocation of IDs for a request. Plain
requests obtain even ID, while interrupt requests are encoded in the low
bit. So, in next patches we will be able to use the rest of ID bits to
calculate hash, and the hash will be the same for plain and interrupt
requests.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>