[fs] fuse: allow using readdir cache
Message-id: <20191029191438.6931-19-vgoyal@redhat.com>
Patchwork-id: 281750
O-Subject: [RHEL8.2 PATCH v2 018/134] fuse: allow using readdir cache
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 5d7bc7e8680c7ca4c8a4f139ce2a54ccb8131ef0
Author: Miklos Szeredi <mszeredi@redhat.com>
Date: Mon, 1 Oct 2018 10:07:04 +0200

The cache is only used if it's completed, not while it's still being
filled; this constraint could be lifted later, if it turns out to be
useful.

Introduce state in struct fuse_file that indicates the position within the
cache.  After a seek, reset the position to the beginning of the cache and
search the cache for the current position.  If the current position is not
found in the cache, then fall back to uncached readdir.

It can also happen that page(s) disappear from the cache, in which case we
must also fall back to uncached readdir.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>