[fs] fuse: use mtime for readdir cache verification
Message-id: <20191029191438.6931-21-vgoyal@redhat.com>
Patchwork-id: 281746
O-Subject: [RHEL8.2 PATCH v2 020/134] fuse: use mtime for readdir cache verification
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 7118883b44b8edfea732aadeb0d4424da3f152b2
Author: Miklos Szeredi <mszeredi@redhat.com>
Date: Mon, 1 Oct 2018 10:07:04 +0200

Store the modification time of the directory in the cache, obtained before
starting to fill the cache.

When reading the cache, verify that the directory hasn't changed, by
checking if current modification time is the same as the one stored in the
cache.

This only needs to be done when the current file position is at the
beginning of the directory, as mandated by POSIX.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>