[fs] fuse: enable caching of symlinks
Message-id: <20191029191438.6931-28-vgoyal@redhat.com>
Patchwork-id: 281681
O-Subject: [RHEL8.2 PATCH v2 027/134] fuse: enable caching of symlinks
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 5571f1e65486be025f73fa6aa30fb03725d362a2
Author: Dan Schatzberg <dschatzberg@fb.com>
Date: Thu, 11 Oct 2018 08:17:00 -0700

FUSE file reads are cached in the page cache, but symlink reads are
not. This patch enables FUSE READLINK operations to be cached which
can improve performance of some FUSE workloads.

In particular, I'm working on a FUSE filesystem for access to source
code and discovered that about a 10% improvement to build times is
achieved with this patch (there are a lot of symlinks in the source
tree).

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>