[fs] fuse: Optimize request_end() by not taking fiq->waitq.lock
Message-id: <20191029191438.6931-41-vgoyal@redhat.com>
Patchwork-id: 281667
O-Subject: [RHEL8.2 PATCH v2 040/134] fuse: Optimize request_end() by not taking fiq->waitq.lock
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 217316a601016d0d2913290070e6338576ae73f6
Author: Kirill Tkhai <ktkhai@virtuozzo.com>
Date: Thu, 8 Nov 2018 12:05:25 +0300

We take global fiq->waitq.lock every time, when we are in this function,
but interrupted requests are just small subset of all requests. This patch
optimizes request_end() and makes it to take the lock when it's really
needed.

queue_interrupt() needs small change for that. After req is linked to
interrupt list, we do smp_mb() and check for FR_FINISHED again. In case of
FR_FINISHED bit has appeared, we remove req and leave the function:

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>