[fs] fuse: add FUSE_WRITE_KILL_PRIV
Message-id: <20191029191438.6931-67-vgoyal@redhat.com>
Patchwork-id: 281698
O-Subject: [RHEL8.2 PATCH v2 066/134] fuse: add FUSE_WRITE_KILL_PRIV
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 4a2abf99f9c2877039006b8dfb3ec4e0d1278fe8
Author: Miklos Szeredi <mszeredi@redhat.com>
Date: Mon, 27 May 2019 09:08:12 +0200

In the FOPEN_DIRECT_IO case the write path doesn't call file_remove_privs()
and that means setuid bit is not cleared if unpriviliged user writes to a
file with setuid bit set.

pjdfstest chmod test 12.t tests this and fails.

Fix this by adding a flag to the FUSE_WRITE message that requests clearing
privileges on the given file.  This needs

This better than just calling fuse_remove_privs(), because the attributes
may not be up to date, so in that case a write may miss clearing the
privileges.

Test case:

  $ passthrough_ll /mnt/pasthrough-mnt -o default_permissions,allow_other,cache=never
  $ mkdir /mnt/pasthrough-mnt/testdir
  $ cd /mnt/pasthrough-mnt/testdir
  $ prove -rv pjdfstests/tests/chmod/12.t

Reported-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Tested-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>