[fs] fuse: extract helper for range writeback
Message-id: <20191029191438.6931-68-vgoyal@redhat.com>
Patchwork-id: 281705
O-Subject: [RHEL8.2 PATCH v2 067/134] fuse: extract helper for range writeback
Bugzilla: 1694161
RH-Acked-by: Miklos Szeredi <mszeredi@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ:1694161

commit 26eb3bae5003585806124c880bba5bef82c80a23
Author: Miklos Szeredi <mszeredi@redhat.com>
Date: Tue, 28 May 2019 13:22:50 +0200

The fuse_writeback_range() helper flushes dirty data to the userspace
filesystem.

When the function returns, the WRITE requests for the data in the given
range have all been completed.  This is not equivalent to fsync() on the
given range, since the userspace filesystem may not yet have the data on
stable storage.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Phillip Lougher <plougher@redhat.com>