[nvme] make fabrics command run on a separate request queue
Message-id: <1572454518-5878-9-git-send-email-dmilburn@redhat.com>
Patchwork-id: 281861
O-Subject: [RHEL8.2 PATCH 8/8] nvme: make fabrics command run on a separate request queue
Bugzilla: 1766361
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>

From: Sagi Grimberg <sagi@grimberg.me>

Bugzilla: 1766361
Build info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=24338630
Test: NetApp ran 12 hours of stress on NVMe over RDMA RoCE configuration.

We have a fundamental issue that fabric commands use the admin_q.
The reason is, that admin-connect, register reads and writes and
admin commands cannot be guaranteed ordering while we are running
controller resets.

For example, when we reset a controller we perform:
1. disable the controller
2. teardown the admin queue
3. re-establish the admin queue
4. enable the controller

In order to perform (3), we need to unquiesce the admin queue, however
we may have some admin commands that are already pending on the
quiesced admin_q and will immediate execute when we unquiesce it before
we execute (4). The host must not send admin commands to the controller
before enabling the controller.

To fix this, we have the fabric commands (admin connect and property
get/set, but not I/O queue connect) use a separate fabrics_q and make
sure to quiesce the admin_q before we disable the controller, and
unquiesce it only after we enable the controller.

This fixes the error prints from nvmet in a controller reset storm test:
kernel: nvmet: got cmd 6 while CC.EN == 0 on qid = 0
Which indicate that the host is sending an admin command when the
controller is not enabled.

Reviewed-by:  James Smart <james.smart@broadcom.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
(cherry picked from commit e7832cb48a654cd12b2bc9181b2f0ad49d526ac6)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>