[netdrv] nfp: flower: encode mac indexes with pre-tunnel rule check
Message-id: <20191031103138.17850-31-pablo.cascon@redhat.com>
Patchwork-id: 282071
Patchwork-instance: patchwork
O-Subject: [RHEL8.2 BZ1729154 30/36] nfp: flower: encode mac indexes with pre-tunnel rule check
Bugzilla: 1729154
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1729154
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=24338621

When a tunnel packet arrives on the NFP card, its destination MAC is
looked up and MAC index returned for it. This index can help verify the
tunnel by, for example, ensuring that the packet arrived on the expected
port. If the packet is destined for a known MAC that is not connected to a
given physical port then the mac index can have a global value (e.g. when
a series of bonded ports shared the same MAC).

If the packet is to be detunneled at a bridge device or internal port like
an Open vSwitch VLAN port, then it should first match a 'pre-tunnel' rule
to direct it to that internal port.

Use the MAC index to indicate if a packet should match a pre-tunnel rule
before decap is allowed. Do this by tracking the number of internal ports
associated with a MAC address and, if the number if >0, set a bit in the
mac_index to forward the packet to the pre-tunnel table before continuing
with decap.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 2e0bc7f3cb5553812f5808ede2cea746aabfbd03)
Signed-off-by: Pablo Cascn <pablo.cascon@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>