[sound] ALSA: firewire-tascam: use the same size of period for PCM substream in AMDTP streams
Message-id: <15728751819462729-51-patchset-jkysela@redhat.com>
Patchwork-id: 282473
O-Subject: [RHEL 8.2 PATCH 050/167] ALSA: firewire-tascam: use the same size of period for PCM substream in AMDTP streams
Bugzilla: 1710702
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

In current implementation, when opening a PCM substream, it's needed to
check whether the opposite PCM substream runs. This is to assign
effectual constraints (e.g. sampling rate) to opened PCM substream.

The number of PCM substreams on AMDTP streams in domain is recorded in
own structure. Usage of this count is an alternative of the above check.
This is better because the count is incremented in pcm.hw_params earlier
than pcm.trigger.

This commit replaces the check with the substream count and the value for
the size of PCM period. Unlike the other drivers in ALSA firewire stack,
no MIDI substream is multiplexed into AMDTP stream.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>
Link: https://lore.kernel.org/r/20191007110532.30270-16-o-takashi@sakamocchi.jp
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 6669a11d35bfa37f319da639594cdcbb164085f4)
Bugzilla: 1710702
Signed-off-by: Phillip Lougher <plougher@redhat.com>