[sound] ALSA: firewire-lib: handle several AMDTP streams in callback handler of IRQ target
Message-id: <15728751819462729-70-patchset-jkysela@redhat.com>
Patchwork-id: 282501
O-Subject: [RHEL 8.2 PATCH 069/167] ALSA: firewire-lib: handle several AMDTP streams in callback handler of IRQ target
Bugzilla: 1710702
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

This commit changes AMDTP domain to run on an IT context of 1394 OHCI as
IRQ target. No hardware interrupt is scheduled for the other isoc
contexts. All of the isoc context are processed in a callback for an isoc
context of IRQ target.

The IRQ target is automatically selected from a list of AMDTP streams,
thus users of AMDTP domain should add an AMDTP stream for IT context
at least.

The reason to select IT context as IRQ target is that the IT context
runs on local 1394 OHCI controller and it can be used as reliable,
constant IRQ generator. On the other hand, IR context can include skip
cycle according to isoc packet transferred by device.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>
Link: https://lore.kernel.org/r/20191018061911.24909-6-o-takashi@sakamocchi.jp
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 60dd49298ec5803ca423836d5a724c3fe402cc3f)
Bugzilla: 1710702
Signed-off-by: Phillip Lougher <plougher@redhat.com>