[kernel] perf/core: Restore mmap record type correctly
Message-id: <20191104222454.10863-30-mpetlan@redhat.com>
Patchwork-id: 282673
O-Subject: [RHEL8.2 BZ1715508 029/327] perf/core: Restore mmap record type correctly
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit d9c1bb2f6a2157b38e8eb63af437cb22701d31ee
Author: Stephane Eranian <eranian@google.com>
Date: Thu Mar 7 10:52:33 2019 -0800

description
===========
On mmap(), perf_events generates a RECORD_MMAP record and then checks
which events are interested in this record. There are currently 2
versions of mmap records: RECORD_MMAP and RECORD_MMAP2. MMAP2 is larger.
The event configuration controls which version the user level tool
accepts.

If the event->attr.mmap2=1 field then MMAP2 record is returned.  The
perf_event_mmap_output() takes care of this. It checks attr->mmap2 and
corrects the record fields before putting it in the sampling buffer of
the event.  At the end the function restores the modified MMAP record
fields.

The problem is that the function restores the size but not the type.
Thus, if a subsequent event only accepts MMAP type, then it would
instead receive an MMAP2 record with a size of MMAP record.

This patch fixes the problem by restoring the record type on exit.

Fixes: 13d7a2410fa6 ("perf: Add attr->mmap2 attribute to an event")
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>