[tools] perf util: Use cached rbtree for rblists
Message-id: <20191104222454.10863-71-mpetlan@redhat.com>
Patchwork-id: 282716
O-Subject: [RHEL8.2 BZ1715508 070/327] perf util: Use cached rbtree for rblists
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit ca2270292e6c3415102242bf9dc3d05f622b7b28
Author: Davidlohr Bueso <dave@stgolabs.net>
Date: Thu Dec 6 11:18:16 2018 -0800

description
===========
At the cost of an extra pointer, we can avoid the O(logN) cost of
finding the first element in the tree (smallest node), which is
something required for any of the strlist or intlist traversals
(XXX_for_each_entry()). There are a number of users in perf of these
(particularly strlists), including probes, and buildid.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>