[tools] perf arm cs-etm: Use event attributes to send sink information to kernel
Message-id: <20191104222454.10863-102-mpetlan@redhat.com>
Patchwork-id: 282741
O-Subject: [RHEL8.2 BZ1715508 101/327] perf arm cs-etm: Use event attributes to send sink information to kernel
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit fa4e819bbca9939c54e87ba994a6a9bffd0247df
Author: Mathieu Poirier <mathieu.poirier@linaro.org>
Date: Thu Jan 31 11:47:12 2019 -0700

description
===========
The communication of sink information for a trace session doesn't work
when more than on CPU is involved in the scenario due to the static
nature of sysfs.  As such communicate the sink information to each event
by using the perf_event::attr:config2 attribute.  The information sent
to the kernel is an hash of the sink's name, which is unique in a
system.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>