[tools] perf report: Move symbol annotation to the resort phase
Message-id: <20191104222454.10863-111-mpetlan@redhat.com>
Patchwork-id: 282755
O-Subject: [RHEL8.2 BZ1715508 110/327] perf report: Move symbol annotation to the resort phase
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit dbd2a1d57f4ee6b909c94072ae9484199308acbd
Author: Jiri Olsa <jolsa@kernel.org>
Date: Mon Feb 4 15:18:08 2019 +0100

description
===========
Currently we make the annotation for the IPC column during the entry
display, already outside of the progress bar scope, so it appears like
'perf report' is stuck.

Move the annotation retrieval to the resort phase, so that all the data
are ready for display.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>