[tools] perf cs-etm: Add exception number in exception packet
Message-id: <20191104222454.10863-116-mpetlan@redhat.com>
Patchwork-id: 282759
O-Subject: [RHEL8.2 BZ1715508 115/327] perf cs-etm: Add exception number in exception packet
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 47106e7413f9fc10780d1ec7df3eaa725df6ebe8
Author: Leo Yan <leo.yan@linaro.org>
Date: Tue Jan 29 20:28:38 2019 +0800

description
===========
When an exception packet comes, it contains the information for
exception number; the exception number indicates the exception types, so
from it we can know if the exception is taken for interrupt, system call
or other traps, etc.

This patch simply adds a field in cs_etm_packet struct, it records
exception number for exception packet that will then be used to properly
identify exception types to the perf synthesize mechanic.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>