[tools] perf cs-etm: Change tuple from traceID-CPU# to traceID-metadata
Message-id: <20191104222454.10863-117-mpetlan@redhat.com>
Patchwork-id: 282761
O-Subject: [RHEL8.2 BZ1715508 116/327] perf cs-etm: Change tuple from traceID-CPU# to traceID-metadata
Bugzilla: 1715508
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715508
https://bugzilla.redhat.com/show_bug.cgi?id=1715508

upstream
========
commit 95c6fe970a0160cb770c5dce9f80311b42d030c0
Author: Leo Yan <leo.yan@linaro.org>
Date: Tue Jan 29 20:28:39 2019 +0800

description
===========
If packet processing wants to know the packet is bound with which ETM
version, it needs to access metadata to decide that based on metadata
magic number; but we cannot simply to use CPU logic ID number as index
to access metadata sequential array, especially when system have
hotplugged off CPUs, the metadata array are only allocated for online
CPUs but not offline CPUs, so the CPU logic number doesn't match with
its index in the array.

This patch is to change tuple from traceID-CPU# to traceID-metadata,
thus it can use the tuple to retrieve metadata pointer according to
traceID.

For safe accessing metadata fields, this patch provides helper function
cs_etm__get_cpu() which is used to return CPU number according to
traceID; cs_etm_decoder__buffer_packet() is the first consumer for this
helper function.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>